# Conflicts:
#	src/main/java/com/djcps/wms/warehouse/controller/AreaController.java
#	src/main/java/com/djcps/wms/warehouse/controller/LocationController.java
#	src/main/java/com/djcps/wms/warehouse/controller/WarehouseController.java
#	src/main/java/com/djcps/wms/warehouse/model/area/AreaCodeBO.java
#	src/main/java/com/djcps/wms/warehouse/model/location/LocationBO.java
#	src/main/java/com/djcps/wms/warehouse/model/warehouse/GetWarehouseCodeBO.java
#	src/main/java/com/djcps/wms/warehouse/server/AreaServer.java
#	src/main/java/com/djcps/wms/warehouse/server/LocationServer.java
#	src/main/java/com/djcps/wms/warehouse/server/WarehouseServer.java
#	src/main/java/com/djcps/wms/warehouse/service/AreaService.java
#	src/main/java/com/djcps/wms/warehouse/service/LocationService.java
#	src/main/java/com/djcps/wms/warehouse/service/WarehouseService.java
#	src/main/java/com/djcps/wms/warehouse/service/impl/AreaServiceImpl.java
#	src/main/java/com/djcps/wms/warehouse/service/impl/LocationServiceImpl.java
#	src/main/java/com/djcps/wms/warehouse/service/impl/WarehouseServiceImpl.java

+++ src/main/java/com/djcps/wms/warehouse/service/impl/AreaServiceImpl.java
@@ -1,37 +1,29 @@
 package com.djcps.wms.warehouse.service.impl;
 
-import java.math.BigDecimal;
-import java.text.DecimalFormat;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 import com.djcps.wms.commons.model.GetCodeBO;
-import com.djcps.wms.commons.model.PartnerInfoBo;
-import com.djcps.wms.warehouse.model.area.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.djcps.wms.address.model.ProvinceCityAreaCodeBo;
+import com.djcps.wms.address.model.ProvinceCityAreaCodeBO;
 import com.djcps.wms.address.server.AddressServer;
-import com.djcps.wms.commons.constant.AppConstant;
+import com.djcps.wms.commons.enums.SysMsgEnum;
 import com.djcps.wms.commons.httpclient.HttpResult;
 import com.djcps.wms.commons.msg.MsgTemplate;
 import com.djcps.wms.warehouse.model.area.AddAreaBO;
-import com.djcps.wms.warehouse.model.area.CountyBo;
-import com.djcps.wms.warehouse.model.area.ProvinceCityBo;
-import com.djcps.wms.warehouse.model.area.SelectAllAreaList;
-import com.djcps.wms.warehouse.model.area.StreetBo;
+import com.djcps.wms.warehouse.model.area.CountyBO;
+import com.djcps.wms.warehouse.model.area.DeleteAreaBO;
+import com.djcps.wms.warehouse.model.area.ProvinceCityBO;
+import com.djcps.wms.warehouse.model.area.SelectAllAreaListBO;
+import com.djcps.wms.warehouse.model.area.StreetBO;
 import com.djcps.wms.warehouse.model.area.UpdateAreaBO;
-import com.djcps.wms.warehouse.model.warehouse.DeleteWarehouseBO;
 import com.djcps.wms.warehouse.model.warehouse.SelectWarehouseByIdBO;
 import com.djcps.wms.warehouse.server.AreaServer;
 import com.djcps.wms.warehouse.service.AreaService;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-import com.mysql.fabric.xmlrpc.base.Array;
-import rpc.plugin.http.HTTPResponse;
 
 /**
  * @title:仓库管理业务层
@@ -62,8 +54,14 @@ public class AreaServiceImpl implements AreaService {
 	 */
 	@Override
 	public Map<String, Object> addArea(AddAreaBO param){
-		HttpResult result = wareAreaServer.addArea(param);
-		return MsgTemplate.customMsg(result);
+		//编码确认
+		HttpResult verifyCode = wareAreaServer.verifyCode(param);
+		if(verifyCode.isSuccess()){
+			HttpResult result = wareAreaServer.addArea(param);
+			return MsgTemplate.customMsg(result);
+		}else{
+			return MsgTemplate.failureMsg(SysMsgEnum.CODE_ERROE);
+		}
 	}
 
 	/**
@@ -91,9 +89,15 @@ public class AreaServiceImpl implements AreaService {
 	 * @date:2017年12月7日
 	 */
 	@Override
-	public Map<String, Object> deleteArea(DeleteWarehouseBO param){
-		HttpResult result = wareAreaServer.deleteArea(param);
-		return MsgTemplate.customMsg(result);
+	public Map<String, Object> deleteArea(DeleteAreaBO param){
+		//删除编码
+		HttpResult deleteCode = wareAreaServer.deleteCode(param);
+		if(deleteCode.isSuccess()){
+			HttpResult result = wareAreaServer.deleteArea(param);
+			return MsgTemplate.customMsg(result);
+		}else{
+			return MsgTemplate.failureMsg(SysMsgEnum.CODE_ERROE);
+		}
 	}
 
 	/**
@@ -106,7 +110,7 @@ public class AreaServiceImpl implements AreaService {
 	 * @date:2017年12月7日
 	 */
 	@Override
-	public Map<String, Object> getAreaAllList(SelectAllAreaList param){
+	public Map<String, Object> getAreaAllList(SelectAllAreaListBO param){
 		HttpResult result = wareAreaServer.getAreaAllList(param);
 		return MsgTemplate.customMsg(result);
 	}
@@ -123,19 +127,19 @@ public class AreaServiceImpl implements AreaService {
 	@Override
 	public Map<String, Object> getAreaById(SelectWarehouseByIdBO param){
 		HttpResult result = wareAreaServer.getAreaById(param);
-		ProvinceCityBo provinceCity = gson.fromJson(gson.toJson(result.getData()), ProvinceCityBo.class);
-		List<CountyBo> countyList = gson.fromJson(gson.toJson(provinceCity.getCountyList()),new TypeToken<List<CountyBo>>(){}.getType()); 
+		ProvinceCityBO provinceCity = gson.fromJson(gson.toJson(result.getData()), ProvinceCityBO.class);
+		List<CountyBO> countyList = gson.fromJson(gson.toJson(provinceCity.getCountyList()),new TypeToken<List<CountyBO>>(){}.getType());
 		List list = provinceCity.getCountyList();
 		if(list.size()!= 0){
-			for (CountyBo countyBo : countyList) {
+			for (CountyBO countyBo : countyList) {
 				//数据库查出来的街道
-				List<StreetBo> sqlStreetList = gson.fromJson(gson.toJson(countyBo.getStreetList()),new TypeToken<List<StreetBo>>(){}.getType());
+				List<StreetBO> sqlStreetList = gson.fromJson(gson.toJson(countyBo.getStreetList()),new TypeToken<List<StreetBO>>(){}.getType());
 				//以下是从地址服务获取的街道
 				String countyCode = countyBo.getCountyCode();
-				ProvinceCityAreaCodeBo code = new ProvinceCityAreaCodeBo();
+				ProvinceCityAreaCodeBO code = new ProvinceCityAreaCodeBO();
 				code.setCode(countyCode);
 				HttpResult streeListByArea = addressServer.getStreeListByArea(code);
-				List<StreetBo> streetList = gson.fromJson(gson.toJson(streeListByArea.getData()),new TypeToken<List<StreetBo>>(){}.getType());
+				List<StreetBO> streetList = gson.fromJson(gson.toJson(streeListByArea.getData()),new TypeToken<List<StreetBO>>(){}.getType());
 				if(sqlStreetList.size() == streetList.size()){
 					countyBo.setCountyStatus("2");
 				}else if(sqlStreetList.size() != streetList.size()){
@@ -143,7 +147,7 @@ public class AreaServiceImpl implements AreaService {
 				}
 			}
 		}else{
-			for (CountyBo countyBo : countyList) {
+			for (CountyBO countyBo : countyList) {
 				countyBo.setCountyStatus("0");
 			}
 		}
