+++ src/main/java/com/djcps/wms/warehouse/service/impl/LocationServiceImpl.java
@@ -1,24 +1,18 @@
 package com.djcps.wms.warehouse.service.impl;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
-import com.djcps.wms.commons.model.GetCodeBO;
-import com.djcps.wms.commons.model.PartnerInfoBo;
+import com.djcps.wms.commons.model.PartnerInfoBO;
 import com.djcps.wms.warehouse.model.location.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.djcps.wms.commons.base.BaseListParam;
+import com.djcps.wms.commons.enums.SysMsgEnum;
 import com.djcps.wms.commons.httpclient.HttpResult;
 import com.djcps.wms.commons.msg.MsgTemplate;
-import com.djcps.wms.warehouse.model.warehouse.AddWarehouseBO;
-import com.djcps.wms.warehouse.model.warehouse.DeleteWarehouseBO;
-import com.djcps.wms.warehouse.model.warehouse.IsUseWarehouseBO;
-import com.djcps.wms.warehouse.model.warehouse.SelectWarehouseByAttributeBO;
-import com.djcps.wms.warehouse.model.warehouse.SelectWarehouseByIdBO;
-import com.djcps.wms.warehouse.model.warehouse.UpdateWarehouseBO;
 import com.djcps.wms.warehouse.server.LocationServer;
-import com.djcps.wms.warehouse.server.WarehouseServer;
 import com.djcps.wms.warehouse.service.LocationService;
 import com.google.gson.Gson;
 
@@ -40,8 +34,14 @@ public class LocationServiceImpl implements LocationService {
 
 	@Override
 	public Map<String, Object> addLocation(AddLocationBO param) {
-		HttpResult result = locationServer.addLocation(param);
-		return MsgTemplate.customMsg(result);
+		//编码确认
+		HttpResult verifyCode = locationServer.verifyCode(param);
+		if(verifyCode.isSuccess()){
+			HttpResult result = locationServer.addLocation(param);
+			return MsgTemplate.customMsg(result);
+		}else{
+			return MsgTemplate.failureMsg(SysMsgEnum.CODE_ERROE);
+		}
 	}
 
 	@Override
@@ -52,12 +52,19 @@ public class LocationServiceImpl implements LocationService {
 
 	@Override
 	public Map<String, Object> deleteLocation(DeleteLocationBO param) {
-		HttpResult result = locationServer.deleteLocation(param);
-		return MsgTemplate.customMsg(result);
+		//删除编码
+		HttpResult deleteCode = locationServer.deleteCode(param);
+		if(deleteCode.isSuccess()){
+			HttpResult result = locationServer.deleteLocation(param);
+			return MsgTemplate.customMsg(result);
+		}else{
+			return MsgTemplate.failureMsg(SysMsgEnum.CODE_ERROE);
+		}
+
 	}
 
 	@Override
-	public Map<String, Object> getLocationAllList(SelectAllLocationList param) {
+	public Map<String, Object> getLocationAllList(SelectAllLocationListBO param) {
 		HttpResult result = locationServer.getLocationAllList(param);
 		return MsgTemplate.customMsg(result);
 	}
@@ -74,5 +81,13 @@ public class LocationServiceImpl implements LocationService {
 		return MsgTemplate.customMsg(result);
 	}
 
-
+	@Override
+	public Map<String, Object> getLocationByCode(SelectLocationByAttributeBO param) {
+        HttpResult result = locationServer.getLocationByAttribute(param);
+        Map map = (Map) result.getData();
+        ArrayList list = (ArrayList) map.get("result");
+        Object object = list.get(0);
+        result.setData(object);
+        return MsgTemplate.customMsg(result);
+    }
 }
